notes and small corrections
