+++ spaceWeatherApi/DataModels/RegionReportModel.cs
@@ -11,4 +11,3 @@
         public int RecentCMECount { get; set; }
     }
 }
-}
