+++ spaceWeatherApi/DataModels/SunspotReportItemModel.cs
@@ -1,0 +1,9 @@
+﻿namespace SpaceWeatherApi.DataModels
+{
+    public class SunspotReportItem
+    {
+        public string? Obsdate { get; set; }
+        public double? NumSpot { get; set; }
+        public string? SpotClass { get; set; }
+    }
+}
